nds: Split various functions into seperate files
