+++ src/main-nds.c
@@ -35,74 +35,10 @@
 #include "ui-init.h"
 
 /* DS includes */
-#include "nds/ds_font_3x8.h"
-#include "nds/ds_main.h"
-
-#define NDS_SCREEN_WIDTH 256
-#define NDS_SCREEN_HEIGHT 192
-
-#define NDS_SCREEN_COLS (NDS_SCREEN_WIDTH / NDS_FONT_WIDTH)
-#define NDS_SCREEN_LINES (NDS_SCREEN_HEIGHT / NDS_FONT_HEIGHT)
-
-#define NDS_BUTTON_FILE "buttons.dat"
-
-#define NDS_MAPPABLE_MASK (KEY_A | KEY_B | KEY_X | KEY_Y | KEY_START | KEY_SELECT)
-#define NDS_MODIFIER_MASK (KEY_L | KEY_R)
-#define NDS_BUTTON_MASK (NDS_MAPPABLE_MASK | NDS_MODIFIER_MASK)
-#define NDS_NUM_MAPPABLE 6 /* A, B, X, Y, Select, Start */
-#define NDS_NUM_MODIFIER 2 /* R, L */
-#define NDS_CMD_LENGTH 16  /* max. 15 keys/button + null terminator */
-
-/* [mappable]*2^[mods] things to map commands to, [cmd_length] chars per command */
-byte nds_btn_cmds[NDS_NUM_MAPPABLE << NDS_NUM_MODIFIER][NDS_CMD_LENGTH];
-
-/* make sure there's something there to start with */
-byte btn_defaults[] = {'a', 'b', 'c', 'd', 'e', 'f', 'g', 'h',
-                       'i', 'j', 'k', 'l', 'm', 'n', 'o', 'p',
-                       'q', 'r', 's', 't', 'u', 'v', 'w', 'z'};
-
-const s16 mappables[] = {KEY_A, KEY_B, KEY_X, KEY_Y, KEY_SELECT, KEY_START};
-const s16 modifiers[] = {KEY_L, KEY_R};
-s16 nds_buttons_to_btnid(u16 kd, u16 kh)
-{
-	if (!(kd & NDS_MAPPABLE_MASK))
-		return -1;
-	u16 i, mods = 0;
-	for (i = 0; i < NDS_NUM_MODIFIER; i++) {
-		if (kh & modifiers[i])
-			mods |= (1 << i);
-	}
-	for (i = 0; i < NDS_NUM_MAPPABLE; i++) {
-		if (kd & mappables[i])
-			return i + NDS_NUM_MAPPABLE * (mods);
-	}
-	return -1;
-}
-
-const nds_kbd_key row0[] = {{16, (u16)'`'}, {16, (u16)'1'},  {16, (u16)'2'},
-                            {16, (u16)'3'}, {16, (u16)'4'},  {16, (u16)'5'},
-                            {16, (u16)'6'}, {16, (u16)'7'},  {16, (u16)'8'},
-                            {16, (u16)'9'}, {16, (u16)'0'},  {16, (u16)'-'},
-                            {16, (u16)'='}, {32, (u16)'\b'}, {0, 0}};
-const nds_kbd_key row1[] = {{24, (u16)'\t'}, {16, (u16)'q'},  {16, (u16)'w'},
-                            {16, (u16)'e'},  {16, (u16)'r'},  {16, (u16)'t'},
-                            {16, (u16)'y'},  {16, (u16)'u'},  {16, (u16)'i'},
-                            {16, (u16)'o'},  {16, (u16)'p'},  {16, (u16)'['},
-                            {16, (u16)']'},  {24, (u16)'\\'}, {0, 0}};
-const nds_kbd_key row2[] = {{32, K_CAPS},    {16, (u16)'a'}, {16, (u16)'s'},
-                            {16, (u16)'d'},  {16, (u16)'f'}, {16, (u16)'g'},
-                            {16, (u16)'h'},  {16, (u16)'j'}, {16, (u16)'k'},
-                            {16, (u16)'l'},  {16, (u16)';'}, {16, (u16)'\''},
-                            {32, (u16)'\n'}, {0, 0}};
-const nds_kbd_key row3[] = {{40, K_SHIFT},  {16, (u16)'z'}, {16, (u16)'x'},
-                            {16, (u16)'c'}, {16, (u16)'v'}, {16, (u16)'b'},
-                            {16, (u16)'n'}, {16, (u16)'m'}, {16, (u16)','},
-                            {16, (u16)'.'}, {16, (u16)'/'}, {40, K_SHIFT},
-                            {0, 0}};
-const nds_kbd_key row4[] = {{32, K_CTRL}, {24, K_ALT},  {128, (u16)' '},
-                            {24, K_ALT},  {32, K_CTRL}, {0, 0}};
-char shifts[] = "`~1!2@3#4$5%6^7&8*9(0)-_=+[{]}\\|;:\'\",<.>/?";
-const nds_kbd_key *kbdrows[] = {row0, row1, row2, row3, row4};
+#include "nds/nds-draw.h"
+#include "nds/nds-event.h"
+#include "nds/nds-keyboard.h"
+#include "nds/nds-buttons.h"
 
 /*
  * Extra data to associate with each "window"
@@ -174,441 +110,31 @@ static void Term_nuke_nds(term *t)
 }
 
 /*
- * Find the square a particular pixel is part of.
- */
-static void pixel_to_square(int *const x, int *const y, const int ox,
-                            const int oy)
-{
-	(*x) = ox / NDS_FONT_WIDTH;
-	(*y) = oy / NDS_FONT_HEIGHT;
-}
-
-/*
  * Handle a touch on the touch screen.
  */
 static void handle_touch(int x, int y, int button, bool press)
 {
 	/* The co-ordinates are only used in Angband format. */
-	pixel_to_square(&x, &y, x, y);
+	nds_pixel_to_square(&x, &y, x, y);
 
 	if (press)
 		Term_mousepress(x, y, button);
-}
-
-/*
- * Touchscreen keyboard handling
- */
-
-static bool shift = false, ctrl = false, alt = false, caps = false;
-
-u16b kbd_mod_code(u16 ret)
-{
-	if (ret & K_MODIFIER)
-		return ret;
-	if (caps && !shift) {
-		if (ret >= 'a' && ret <= 'z')
-			ret -= 0x20;
-	}
-	if (shift) {
-		char *temp;
-		if (!caps && ret >= 'a' && ret <= 'z')
-			ret -= 0x20;
-		if ((temp = strchr(shifts, ret)) != NULL)
-			ret = *(temp + 1);
-	}
-	if (alt) {
-		ret |= 0x80;
-	}
-	if (ctrl /* && ret >= 'a' && ret < 'a'+32*/) {
-		ret = ret & 0x1f;
-	}
-	return ret;
-}
-
-void kbd_set_color_from_pos(u16b r, u16b k, byte color)
-{
-	u16b ii, xx = 0, jj;
-	u16b *map[] = {(u16b *)(BG_MAP_RAM_SUB(8) + 3 * 32 * 2),
-	               (u16b *)(BG_MAP_RAM_SUB(9) + 3 * 32 * 2),
-	               (u16b *)(BG_MAP_RAM_SUB(10) + 3 * 32 * 2),
-	               (u16b *)(BG_MAP_RAM_SUB(11) + 3 * 32 * 2)};
-	for (ii = 0; ii < k; ii++) {
-		xx += kbdrows[r][ii].width >> 3;
-	}
-	for (ii = 0; ii < (kbdrows[r][k].width >> 3); ii++) {
-		for (jj = 0; jj < 4; jj++) {
-			map[jj][(7 + r * 2) * 32 + ii + xx + 1] =
-			    (map[jj][(7 + r * 2) * 32 + ii + xx + 1] &
-			     0x0FFF) |
-			    (color << 12);
-			map[jj][(7 + r * 2 + 1) * 32 + ii + xx + 1] =
-			    (map[jj][(7 + r * 2 + 1) * 32 + ii + xx + 1] &
-			     0x0FFF) |
-			    (color << 12);
-		}
-	}
-}
-
-void kbd_set_color_from_code(u16b code, byte color)
-{
-	u16b r, k;
-	for (r = 0; r < 5; r++) {
-		for (k = 0; kbdrows[r][k].width != 0; k++) {
-			if (kbd_mod_code(kbdrows[r][k].code) == code) {
-				kbd_set_color_from_pos(r, k, color);
-			}
-			/* 
-			 * do not break!! there may be >1 key with this code
-			 * (modifier keys)
-			 */
-		}
-	}
-}
-
-void kbd_set_map()
-{
-	REG_BG0CNT_SUB = BG_TILE_BASE(0) |
-	                 BG_MAP_BASE(8 + (caps | (shift << 1))) |
-	                 BG_PRIORITY(0) | BG_COLOR_16;
-}
-
-u16b kbd_xy2key(byte x, byte y)
-{
-	/* on arrow-pad */
-	if (x >= 104 && x < 152 && y >= 24 && y < 72) {
-		byte kx = (x - 104) / 16, ky = (y - 24) / 16;
-		return (kx + (2 - ky) * 3 + 1 +
-		        '0') /* | (shift ? K_SHIFTED_MOVE : 0)*/;
-	}
-
-	if (y >= 80 && y < 96) {
-		if (x >= 8 && x < 24)
-			return '\033';
-
-		/* F-key */
-		if (x >= 40 && x < 248) {
-			x -= 40;
-			y = x / 72;  /* which section */
-			x -= y * 72; /* offset in section */
-			if (x < 64) {
-				/* section*4 + offset/16 + 1 */
-				return K_F(y * 4 + (x >> 4) + 1); 
-			} else {
-				return 0;
-			}
-		}
-	}
-
-	s16b ox = x - 8, oy = y - 104;
-	if (ox < 0 || ox >= 240)
-		return 0;
-	if (oy < 0 || oy >= 80)
-		return 0;
-	u16b row = oy / 16;
-	int i;
-	for (i = 0; ox > 0; ox -= kbdrows[row][i++].width)
-		;
-	u16b ret = kbdrows[row][i - 1].code;
-	return kbd_mod_code(ret);
-}
-
-void kbd_dotoggle(bool *flag, int how)
-{
-	switch (how) {
-	case 0:
-		*flag = false;
-		return;
-	case 1:
-		*flag = true;
-		return;
-	default:
-	case -1:
-		*flag = !*flag;
-		return;
-	}
-}
-
-/*
- * which: K_SHIFT, K_CTRL, K_ALT, K_MODIFIER=all keys
- * how: -1 = toggle, 0 = off, 1 = on
- */
-void kbd_togglemod(int which, int how)
-{
-	/* boolean old_shift = shift, old_ctrl = ctrl, old_alt = alt, old_caps = caps; */
-	switch (which) {
-	case K_CTRL:
-		kbd_dotoggle(&ctrl, how);
-		break;
-	case K_SHIFT:
-		kbd_dotoggle(&shift, how);
-		break;
-	case K_ALT:
-		kbd_dotoggle(&alt, how);
-		break;
-	case K_CAPS:
-		kbd_dotoggle(&caps, how);
-		break;
-	case K_MODIFIER:
-		kbd_dotoggle(&ctrl, how);
-		kbd_dotoggle(&shift, how);
-		kbd_dotoggle(&alt, how);
-		/*
-		 * NOT caps!!  This is called to un-set shift, ctrl, and alt after
-		 * a key is pressed.  Unsetting caps here would cause it to be the
-		 * same as shift.
-		 */
-		break;
-	}
-
-	/* if (old_shift != shift) */
-	kbd_set_color_from_code(K_SHIFT, shift);
-
-	/* if (old_ctrl != ctrl) */
-	kbd_set_color_from_code(K_CTRL, ctrl);
-
-	/* if (old_alt != alt) */
-	kbd_set_color_from_code(K_ALT, alt);
-
-	/* if (old_caps != caps) */
-	kbd_set_color_from_code(K_CAPS, caps);
-
-	kbd_set_map();
-}
-
-/*
- * clear this to prevent alt-b, f5, and f6 from having their special effects
- * it's cleared during getlin, yn_function, etc
- */
-byte process_special_keystrokes = 1;
-
-/* 
- * run this every frame
- * returns a key code if one has been typed, else returns 0
- * assumes scankeys() was already called this frame (in real vblank handler)
- */
-byte kbd_vblank()
-{
-	/* frames the stylus has been held down for */
-	static u16b touched = 0;
-	/* coordinates from each frame, the median is used to get the keycode */
-	static s16b xarr[3], yarr[3];
-	/* the keycode of the last key pressed, so it can be un-highlighted */
-	static u16b last_code;
-	/* the keycode of the currently pressed key, is usu. returned */
-	u16b keycode;
-	/* current input data */
-	touchPosition touch;
-
-	touchRead(&touch);
-
-	/* if screen is being touched... */
-	if (keysHeld() & KEY_TOUCH) {
-		if (touched < 3) { /* if counter < 3... */
-			touched++; /* add to counter */
-			xarr[touched - 1] =
-			    touch.px; /* add this to the array for */
-			yarr[touched - 1] = touch.py; /* finding the median */
-		}
-	} else {             /* not being touched */
-		touched = 0; /* so reset the counter for next time */
-	}
-
-	/* if the stylus was released */
-	if (keysUp() & KEY_TOUCH) {
-		/* if last_code is set and it wasn't a modifier */
-		if (last_code && !(last_code & K_MODIFIER)) {
-			/* clear the hiliting on this key */
-			kbd_set_color_from_code(last_code, 0);
-			/* and also clear all modifiers (except caps)    */
-			kbd_togglemod(K_MODIFIER, 0);
-		}
-		last_code = 0;
-	}
-
-	/* if the screen has been touched for 3 frames... */
-	if (touched == 3) {
-		touched++; /* do not return the keycode again */
-		/* also, not setting to zero prevents the keysHeld() thing */
-		/*  from starting the process over and getting 3 more samples */
-
-		u16b i, tmp, the_x = 0, the_y = 0;
-
-		/* x/yarr now contains 3 values from each of the 3 frames */
-		/* take the median of each array and put into the_x/y */
-
-		/* sort the array */
-		/* bubble sort, ugh */
-		for (i = 1; i < 3; i++) {
-			if (xarr[i] < xarr[i - 1]) {
-				tmp = xarr[i];
-				xarr[i] = xarr[i - 1];
-				xarr[i - 1] = tmp;
-			}
-			if (yarr[i] < yarr[i - 1]) {
-				tmp = yarr[i];
-				yarr[i] = yarr[i - 1];
-				yarr[i - 1] = tmp;
-			}
-		}
-
-		/* get the middle value (median) */
-		/* if it's -1, take the top value */
-		if (xarr[1] == -1)
-			the_x = xarr[2];
-		else
-			the_x = xarr[1];
-		if (yarr[1] == -1)
-			the_y = yarr[2];
-		else
-			the_y = yarr[1];
-
-		/* get the keycode that corresponds to this key */
-		keycode = kbd_xy2key(the_x, the_y);
-
-		/* if it's not a modifier, highlight it */
-		if (keycode && !(keycode & K_MODIFIER))
-			kbd_set_color_from_code(keycode, 1);
-		/* set last_code so it can be un-highlighted later */
-		last_code = keycode;
-
-		/*/* check for special keystrokes: alt-b, f5, f6 */
-		if (process_special_keystrokes) {
-			/* alt-b: assign button macro */
-			if (keycode == ('b' | 0x80)) {
-				/* clear hiliting */
-				kbd_set_color_from_code(keycode, 0);
-				kbd_togglemod(K_MODIFIER, 0);
-				// nds_assign_button();
-				keycode = last_code =
-				    0; /* don't let nethack process it */
-			}
-
-			if (keycode & K_F(0)) { /* its an f-key */
-				kbd_togglemod(K_MODIFIER, 0);
-			}
-		}
-
-		/* if it's a modifier, toggle it */
-		if (keycode & K_MODIFIER)
-			kbd_togglemod(keycode, -1);
-		else if ((keycode & 0x7F) !=
-		         0) { /* it's an actual keystroke, return it */
-			return (keycode & 0xFF);
-		}
-	}
-
-	return 0;
-}
-
-void nds_check_buttons(u16b kd, u16b kh)
-{
-	s16b btn = nds_buttons_to_btnid(kd, kh);
-	if (btn == -1)
-		return;
-	byte *cmd = &nds_btn_cmds[btn][0];
-	while (*cmd != 0) {
-		put_key_event(*(cmd++));
-	}
-}
-
-/*
- * All event handling
- */
-u16b *ebuf = (u16b *)(&BG_GFX[256 * 192]);
-/* store the queue just past mainscreen display data */
-u16b ebuf_read = 0, ebuf_write = 0;
-byte nds_updated = 0; /* windows that have been updated and should be redrawn */
-
-bool has_event()
-{
-	return ((ebuf[ebuf_read] & EVENT_SET) || (ebuf_read < ebuf_write));
-	/* read < write should never happen without EVENT_SET, but */
-	/* just in case... */
-}
-
-u16b get_event()
-{
-	if (!has_event())
-		return 0;
-	u16b r = ebuf[ebuf_read];
-	ebuf[ebuf_read] = 0;
-	ebuf_read++;
-	if (ebuf_read > ebuf_write) {
-		ebuf_write++;
-		if (ebuf_write >= MAX_EBUF)
-			ebuf_write = 0;
-	}
-	if (ebuf_read >= MAX_EBUF)
-		ebuf_read = 0;
-	return r;
-}
-
-void put_key_event(byte c)
-{
-	ebuf[ebuf_write++] = EVENT_SET | (u16)c;
-	if (ebuf_write >= MAX_EBUF)
-		ebuf_write = 0;
-}
-
-void put_mouse_event(byte x, byte y)
-{
-	ebuf[ebuf_write++] =
-	    EVENT_SET | MEVENT_FLAG | (u16b)x | (((u16b)y) << 7);
-	if (ebuf_write >= MAX_EBUF)
-		ebuf_write = 0;
 }
 
 void do_vblank()
 {
 	swiWaitForVBlank();
 
-	/* --------------------------- */
-	/*  Handle the arrow buttons */
+	/* Check hardware for new inputs */
 	scanKeys();
-	u32b kd = keysDown();
-	u32b kh = keysHeld();
-	/* order of keys: Right, Left, Up, Down */
-	/* map keys to dirs, depends on order of keys in nds/input.h */
-	/*  and order of directions in ndir & sdir in decl.c */
-	/*const s8 k2d[] = {	// indexes into ndir/sdir, 10 = end of string
-	 * = '\0' */
-	/* 10, 4, 0, 10, 2, 3, 1, 10, 6, 5, 7	// no working combinations >=
-	 * 11 */
-	/*}; */
-	const byte k2d[] = {'6', '4', '8', '2', '3', '7', '9', '1'};
-	/* only do stuff if a key was pressed last frame */
-	if (kd & (KEY_RIGHT | KEY_LEFT | KEY_UP | KEY_DOWN)) {
-		u16b dirs_down = 0;
-		int i;
-		if (kh & KEY_LEFT)
-			dirs_down++;
-		if (kh & KEY_RIGHT)
-			dirs_down++;
-		if (kh & KEY_UP)
-			dirs_down++;
-		if (kh & KEY_DOWN)
-			dirs_down++;
-		if (dirs_down == 1 && !(kh & (KEY_R | KEY_L))) {
-			for (i = 0; i < 4; i++)
-				if (kh & (1 << (i + 4)))
-					put_key_event(k2d[i]);
-		} else if (dirs_down == 2 && (kh & (KEY_R | KEY_L))) {
-			for (i = 0; i < 4; i++)
-				if (kh & (1 << (i + 4)))
-					put_key_event(k2d[i + 4]);
-		}
-	}
+	
+	/* Handle button inputs */
+	nds_btn_vblank();
 
-	/* --------------------------- */
-	/*  Check for button macros */
-	nds_check_buttons(kd, kh);
-
-	/* --------------------------- */
-	/*  Check for typing on the touchscreen kbd */
-	byte keycode = kbd_vblank();
+	/* Handle touchscreen (keyboard) inputs */
+	byte keycode = nds_kbd_vblank();
 	if ((keycode & 0x7F) != 0) { /* it's an actual keystroke, return it */
-		put_key_event(keycode & 0xFF);
-		/*Term_keypress(keycode & 0xFF); */
+		nds_event_put_key(keycode & 0xFF);
 	}
 }
 
@@ -630,11 +156,11 @@ static errr CheckEvents(bool wait)
 
 	do_vblank();
 
-	if (!wait && !has_event())
+	if (!wait && !nds_event_ready())
 		return (1);
 
 	while (!e) {
-		e = get_event();
+		e = nds_event_get();
 
 		do_vblank();
 	}
@@ -700,7 +226,7 @@ static errr Term_xtra_nds(int n, int v)
 
 		for (y = 0; y < NDS_SCREEN_LINES; y++) {
 			for (x = 0; x < NDS_SCREEN_COLS; x++) {
-				draw_char(x, y, 0);
+				nds_draw_char(x, y, 0);
 			}
 		}
 
@@ -839,41 +365,6 @@ static errr Term_curs_nds(int x, int y)
 	return (0);
 }
 
-void draw_char(byte x, byte y, char c)
-{
-	u32b vram_offset = (y & 0x1F) * NDS_FONT_HEIGHT * NDS_SCREEN_WIDTH + x * NDS_FONT_WIDTH;
-
-	u16b *fb = BG_GFX;
-	if (y & 32) {
-		fb = &BG_GFX_SUB[16 * 1024];
-	}
-
-	byte xx, yy;
-	for (yy = 0; yy < NDS_FONT_HEIGHT; yy++) {
-		for (xx = 0; xx < NDS_FONT_WIDTH; xx++) {
-			fb[yy * NDS_SCREEN_WIDTH + xx + vram_offset] = nds_font_pixel(c, xx, yy) | BIT(15);
-		}
-	}
-}
-
-void draw_color_char(byte x, byte y, char c, byte clr)
-{
-	u32b vram_offset = (y & 0x1F) * NDS_FONT_HEIGHT * NDS_SCREEN_WIDTH + x * NDS_FONT_WIDTH;
-
-	u16b *fb = BG_GFX;
-	if (y & 32) {
-		fb = &BG_GFX_SUB[16 * 1024];
-	}
-
-	byte xx, yy;
-	u16b fgc = color_data[clr & BASIC_COLORS];
-	for (yy = 0; yy < NDS_FONT_HEIGHT; yy++) {
-		for (xx = 0; xx < NDS_FONT_WIDTH; xx++) {
-			fb[yy * NDS_SCREEN_WIDTH + xx + vram_offset] = (nds_font_pixel(c, xx, yy) & fgc) | BIT(15);
-		}
-	}
-}
-
 /*
  * Erase some characters
  *
@@ -889,7 +380,7 @@ static errr Term_wipe_nds(int x, int y, int n)
 
 	/* Draw a blank */
 	for (i = 0; i < n; i++)
-		draw_color_char(x + i, y, 0, 0);
+		nds_draw_color_char(x + i, y, 0, 0);
 
 	/* Success */
 	return (0);
@@ -930,7 +421,7 @@ static errr Term_wipe_nds(int x, int y, int n)
 static errr Term_text_nds(int x, int y, int n, int a, const wchar_t *s)
 {
 	for (int i = 0; i < n; i++) {
-		draw_color_char(x + i, y, s[i], a);
+		nds_draw_color_char(x + i, y, s[i], color_data[a & BASIC_COLORS]);
 	}
 
 	return (0);
@@ -1068,126 +559,6 @@ static void init_stuff(void)
 	// strcpy(savefile, "/angband/lib/save/PLAYER");
 }
 
-void nds_log(const char *msg)
-{
-	static byte x = 2, y = 1;
-	byte i = 0;
-	for (i = 0; msg[i] != '\0'; i++) {
-		draw_char(x, y, msg[i]);
-		x++;
-		if (msg[i] == '\n' || x > NDS_SCREEN_COLS - 2) {
-			x = 2;
-			y++;
-		}
-	}
-}
-
-void nds_logf(const char* format, ...)
-{
-	va_list args;
-	va_start(args, format);
-
-	int len = vsnprintf(NULL, 0, format, args);
-
-	char buf[len];
-
-	vsprintf(buf, format, args);
-	nds_log(buf);
-
-	va_end(args);
-}
-
-bool nds_load_kbd()
-{
-	const char *files[] = {
-	    "/angband/nds/kbd.bin",
-	    "/angband/nds/kbd.pal",
-	    "/angband/nds/kbd.map",
-	};
-	char *dests[] = {
-	    (char *) BG_TILE_RAM_SUB(0),
-	    (char *) BG_PALETTE_SUB,
-	    (char *) BG_MAP_RAM_SUB(8),
-	};
-
-	for (int i = 0; i < N_ELEMENTS(files); i++) {
-		ang_file *handle = file_open(files[i], MODE_READ, -1);
-
-		if (!handle) {
-			nds_logf("Error opening '%s'\n", files[i]);
-			return false;
-		}
-
-		char *dest = dests[i];
-		int read_bytes;
-
-		while ((read_bytes = file_read(handle, dest, 1024))) {
-			if (read_bytes == -1) {
-				nds_logf("Error reading '%s'\n", files[i]);
-				return false;
-			}
-
-			if (read_bytes != 1024) {
-				/* We're done reading the file */
-				break;
-			}
-
-			dest += read_bytes;
-		}
-
-		file_close(handle);
-	}
-
-	return true;
-}
-
-void kbd_init()
-{
-	u16b i;
-	for (i = 0; i < 16; i++) {
-		BG_PALETTE_SUB[i + 16] = BG_PALETTE_SUB[i] ^ 0x7FFF;
-	}
-}
-
-void nds_init_buttons()
-{
-	u16b i, j;
-	for (i = 0; i < (NDS_NUM_MAPPABLE << NDS_NUM_MODIFIER); i++) {
-		for (j = 0; j < NDS_CMD_LENGTH; j++) {
-			nds_btn_cmds[i][j] = 0;
-		}
-	}
-	if (access(NDS_BUTTON_FILE, 0444) == -1) {
-		/* Set defaults */
-		for (i = 0; i < (NDS_NUM_MAPPABLE << NDS_NUM_MODIFIER); i++)
-			nds_btn_cmds[i][0] = btn_defaults[i];
-
-		return;
-	}
-
-	FILE *f = fopen(NDS_BUTTON_FILE, "r");
-	fread(&nds_btn_cmds[0], NDS_CMD_LENGTH,
-	      (NDS_NUM_MAPPABLE << NDS_NUM_MODIFIER), f);
-	fclose(f);
-}
-
-void nds_raw_print(const char *str)
-{
-	static u16b x = 0, y = 32;
-	while (*str) {
-		draw_char(x, y, (u8)(*(str++)));
-		x++;
-		if (x > 78) {
-			x = 0;
-			y++;
-			if (y > 34)
-				y = 32;
-		}
-	}
-	draw_char(x, y, 219);
-	fflush(0);
-}
-
 /*
  * Display warning message (see "z-util.c")
  */
@@ -1203,7 +574,6 @@ void nds_exit(int code)
 {
 	u16b i;
 	for (i = 0; i < 60; i++) {
-		nds_updated = 0xFF;
 		do_vblank(); /* wait 1 sec. */
 	}
 	systemShutDown();
@@ -1287,7 +657,7 @@ int main(int argc, char *argv[])
 	swiWaitForVBlank();
 
 	chdir("/angband");
-	if (!nds_load_kbd()) {
+	if (!nds_kbd_init()) {
 		nds_log("\nError loading keyboard graphics.\nCannot continue.\n");
 
 		/* Lock up */
@@ -1296,8 +666,7 @@ int main(int argc, char *argv[])
 
 		return 1; /* die */
 	}
-	kbd_init();
-	nds_init_buttons();
+	nds_btn_init();
 
 	/* Activate hooks */
 	plog_aux = hook_plog;
