--- src/nds/ds_main.h
@@ -1,16 +1,0 @@
-
-#ifndef DS_MAIN_H
-#define DS_MAIN_H
-
-#include <nds.h>
-#include "ds_gfx.h"
-#include "ds_win.h"
-#include "ds_io.h"
-#include "ds_kbd.h"
-#include "ds_btn.h"
-
-void do_vblank();
-void nds_debug_print();
-
-#endif
-
