+++ src/nds/nds-font-3x8.h
@@ -1,10 +1,10 @@
-#ifndef DS_FONT_3X8_H
-#define DS_FONT_3X8_H
+#ifndef _NDS_FONT_3X8_H
+#define _NDS_FONT_3X8_H
 
 #define NDS_FONT_WIDTH 3
 #define NDS_FONT_HEIGHT 8
 
-const u16 ds_subfont[] = {
+static const u16 ds_subfont[] = {
 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 
 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x001f, 0xffff, 0xffe0, 0xffe0, 0x0000, 0x001f, 0xffff, 0x7c00, 
 0xffff, 0xffe0, 0x0000, 0x001f, 0xffe0, 0xffff, 0x7c1f, 0xffe0, 0xffff, 0x001f, 0xffe0, 0x0000, 0x001f, 0x001f, 0xffff, 0xffe0, 
@@ -392,7 +392,7 @@ const u16 ds_subfont[] = {
 };
 
 static inline u16 nds_font_pixel(char c, char subX, char subY) {
-	u32b char_offset = c * NDS_FONT_HEIGHT * NDS_FONT_WIDTH;
+	u32 char_offset = c * NDS_FONT_HEIGHT * NDS_FONT_WIDTH;
 	return ds_subfont[char_offset + subY * NDS_FONT_WIDTH + subX];
 }
